pyproject.toml
[project]
name = "ll-edge-brain"
version = "0.1.0"
description = "Self-contained offline edge RAG brain"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "fastembed>=0.4.0",
    "qdrant-edge-py>=0.6.0",
    "llama-cpp-python>=0.3.0",
    "liteparse>=1.2.0",
    "langdetect>=1.0.9",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
